refactor(colorize): use class to encapsulate global state
